Fixed bug in reading less than file and page:
When there are more bytes in the file left, than there are being read,
but the number of bytes left is less than the number of pages,
the end of the file would not be reached, and a whole page would be read,
causing an error.

Fixed bug by reading only up to the end of the file or a full page,
regardless of the number of bytes desired.

Added test, "smaller_read", for the fix.